Refactor navigation components
Move code to subdir
Extract CSS
Add component extension to component files
