+++ src/js/index.js
@@ -11,9 +11,9 @@ import './components/footer';
 import './components/header';
 import './components/hero';
 import './components/imagefeature';
-import './components/nav';
-import './components/nav-link';
-import './components/nav-section';
+import './components/navigation/navigation.component';
+import './components/navigation/link.component';
+import './components/navigation/section.component';
 import './components/pagination/pagination.component';
 import './components/search';
 import './components/section-nav';
